misc: reduce casts

+++ fyi_witcher/src/witching.rs
@@ -142,7 +142,7 @@ const MIN_DRAW_WIDTH: usize = 40;
 struct WitchingInner {
 	buf: Vec<u8>,
 	toc: Toc,
-	elapsed: u32,
+	elapsed: u64,
 	last_hash: u64,
 	last_lines: usize,
 	last_time: u128,
@@ -260,8 +260,8 @@ impl WitchingInner {
 	/// # Elapsed (Seconds).
 	///
 	/// Return the elapsed time in seconds.
-	pub(crate) fn elapsed(&self) -> u32 {
-		86400.min(self.started.elapsed().as_secs() as u32)
+	pub(crate) fn elapsed(&self) -> u64 {
+		86400.min(self.started.elapsed().as_secs())
 	}
 
 	/// # Percent.
@@ -637,7 +637,7 @@ impl WitchingInner {
 	/// A value of `true` is returned if one or more seconds has elapsed since
 	/// the last tick, otherwise `false` is returned.
 	fn tick_set_secs(&mut self) -> bool {
-		let secs: u32 = self.elapsed();
+		let secs: u64 = self.elapsed();
 		if secs == self.elapsed { false }
 		else {
 			self.elapsed = secs;
@@ -646,7 +646,7 @@ impl WitchingInner {
 				self.toc.replace(&mut self.buf, PART_ELAPSED, b"23:59:59");
 			}
 			else {
-				let c = utility::secs_chunks(secs);
+				let c = utility::secs_chunks(secs as u32);
 				let rgs: usize = self.toc.start(PART_ELAPSED);
 				self.buf[rgs..rgs + 2].copy_from_slice(time_format_dd(c[0]));
 				self.buf[rgs + 3..rgs + 5].copy_from_slice(time_format_dd(c[1]));
@@ -1087,7 +1087,7 @@ impl Witching {
 	///     X files in M minutes and S seconds.
 	fn summary(&self) -> Msg {
 		Msg::from([
-			NiceInt::from(u64::from(self.total())).as_bytes(),
+			NiceInt::from(self.total()).as_bytes(),
 			b" ",
 			self.label(),
 			b" in ",
@@ -1166,7 +1166,7 @@ impl Witching {
 	// These just return the inner values.
 	get_inner!(doing, u32);
 	get_inner!(done, u32);
-	get_inner!(elapsed, u32);
+	get_inner!(elapsed, u64);
 	get_inner!(percent, f64);
 	get_inner!(total, u32);
 	get_inner!(is_running, bool);
